requirements.txt
camelcase==0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
greenlet==1.1.3
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
msgpack==1.0.4
mysql-connector-python==8.0.30
neovim==0.3.1
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.2.0
protobuf==3.20.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynvim==0.4.3
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==1.5
python-decouple==3.6
python-dotenv==0.21.0
python-http-client==3.3.7
pytweening==1.0.4
requests==2.28.1
sendgrid==6.9.7
starkbank-ecdsa==2.1.0
urllib3==1.26.12
Werkzeug==2.2.2
